// src/shared/api/generated.ts
export type components = {
	schemas: {
// ...
		FlavorResponse: {
			/** The body of the resource. */
			body?: components["schemas"]["FlavorResponseBody"] | null;
			/** The metadata related to this resource. */
			metadata?: components["schemas"]["FlavorResponseMetadata"] | null;
			/** The resources related to this resource. */
			resources?: components["schemas"]["FlavorResponseResources"] | null;
			/**
			 * The unique resource id.
			 * Format: uuid
			 */
			id: string;
			/**
			 * Permission Denied
			 * @default false
			 */
			permission_denied: boolean;
			/** The name of the Flavor. */
			name: string;
		};
// ...
		OAuthTokenResponse: {
			/** Access Token */
			access_token: string;
// ...
		};
// ...
		UserResponse: {
			/** The body of the resource. */
			body?: components["schemas"]["UserResponseBody"] | null;
			/** The metadata related to this resource. */
			metadata?: components["schemas"]["UserResponseMetadata"] | null;
			/** The resources related to this resource. */
			resources?: components["schemas"]["UserResponseResources"] | null;
			/**
			 * The unique resource id.
			 * Format: uuid
			 */
			id: string;
			/**
			 * Permission Denied
			 * @default false
			 */
			permission_denied: boolean;
			/** The unique username for the account. */
			name: string;
		};
// ...
	};
	responses: never;
	parameters: never;
	requestBodies: never;
	headers: never;
	pathItems: never;
};